Modify : 피드창 컴포넌트화로 변경
