+++ src/pages/LeeSoo/Main/LeeSooMain.scss
@@ -22,139 +22,135 @@ body {
     height: auto;
     background-color: #f8f4f4;
 
-    .contents {
-      max-width: 700px;
-      max-height: 800px;
-      width: 100%;
-      height: 100%;
-      padding: 30px 10px;
+    .Feed-list {
+      .contents {
+        padding: 30px 10px;
 
-      article {
-        max-width: 614px;
-        height: auto;
-        margin-bottom: 50px;
-        border: 1px solid rgba(45, 42, 42, 0.3);
-        background-color: white;
-
-        .article-top {
-          display: flex;
-          align-items: center;
-          position: relative;
-          padding: 10px;
+        article {
+          max-width: 630px;
+          width: 100%;
+          height: auto;
+          margin-bottom: 50px;
+          border: 1px solid rgba(45, 42, 42, 0.3);
           background-color: white;
 
-          .dotMore {
-            position: absolute;
-            right: 4%;
-          }
+          .article-top {
+            display: flex;
+            align-items: center;
+            position: relative;
+            padding: 10px;
+            background-color: white;
 
-          p {
-            margin-left: 20px;
-            font-weight: 550;
-          }
-
-          .userImg {
-            width: 54px;
-            height: 54px;
-            margin-left: 20px;
-            border-radius: 50%;
-            object-fit: cover;
-            border: 3px solid rgb(211, 55, 125);
-            padding: 1px;
-          }
-        }
-
-        .article-main {
-          max-width: 612px;
-          max-height: auto;
-
-          img {
-            width: 100%;
-            height: 612px;
-            object-fit: cover;
-          }
-        }
-
-        .article-bar {
-          padding: 20px;
-          background-color: white;
-
-          button {
-            border: none;
-            background: none;
-            cursor: pointer;
-          }
-
-          span {
-            color: rgba(19, 18, 18, 0.3);
-            transition: all 1s;
-          }
-
-          .bar-left {
-            span {
-              margin-right: 20px;
+            .dotMore {
+              position: absolute;
+              right: 4%;
             }
 
-            display: flex;
-            float: left;
+            p {
+              margin-left: 20px;
+              font-weight: 550;
+            }
+
+            .userImg {
+              width: 54px;
+              height: 54px;
+              margin-left: 20px;
+              border-radius: 50%;
+              object-fit: cover;
+              border: 3px solid rgb(211, 55, 125);
+              padding: 1px;
+            }
           }
 
-          .bar-like:hover {
-            color: red;
-            transition: all 1s;
+          .article-main {
+            img {
+              width: 100%;
+              height: auto;
+              object-fit: cover;
+            }
           }
 
-          .bar-comments:hover {
-            color: black;
-            transition: all 1s;
-          }
+          .article-bar {
+            padding: 20px;
+            background-color: white;
 
-          .bar-share:hover {
-            color: black;
-            transition: all 1s;
-          }
+            button {
+              border: none;
+              background: none;
+              cursor: pointer;
+            }
 
-          .bar-right {
-            float: right;
+            span {
+              color: rgba(19, 18, 18, 0.3);
+              transition: all 1s;
+            }
 
-            .bar-bookmark:hover {
+            .bar-left {
+              span {
+                margin-right: 20px;
+              }
+
+              display: flex;
+              float: left;
+            }
+
+            .bar-like:hover {
+              color: red;
+              transition: all 1s;
+            }
+
+            .bar-comments:hover {
               color: black;
               transition: all 1s;
             }
-          }
 
-          .clearFloat {
-            clear: both;
+            .bar-share:hover {
+              color: black;
+              transition: all 1s;
+            }
+
+            .bar-right {
+              float: right;
+
+              .bar-bookmark:hover {
+                color: black;
+                transition: all 1s;
+              }
+            }
+
+            .clearFloat {
+              clear: both;
+            }
           }
         }
-      }
 
-      .Comments-show {
-        margin-top: 10px;
-        height: 90px;
-        overflow-y: scroll;
+        .Comments-show {
+          margin-top: 10px;
+          height: 90px;
+          overflow-y: scroll;
 
-        .Comments-list {
-          margin-top: 5px;
-          position: relative;
-          padding: 5px 50px;
-          background-color: white;
+          .Comments-list {
+            margin-top: 5px;
+            position: relative;
+            padding: 5px 50px;
+            background-color: white;
 
-          .nick {
-            font-weight: bold;
-          }
+            .nick {
+              font-weight: bold;
+            }
 
-          .delete {
-            position: absolute;
-            right: 10px;
-            color: gray;
-            cursor: pointer;
-          }
+            .delete {
+              position: absolute;
+              right: 10px;
+              color: gray;
+              cursor: pointer;
+            }
 
-          .like {
-            position: absolute;
-            right: 40px;
-            cursor: pointer;
+            .like {
+              position: absolute;
+              right: 40px;
+              cursor: pointer;
+            }
           }
         }
       }
